src/server/requirements.txt
asgiref==3.4.1
black==22.8.0
click==8.0.4
dataclasses==0.8
Django==3.0.3
django-cors-headers==3.10.1
django-environ==0.9.0
django-paranoid==1.0
djangorestframework==3.14.0
importlib-metadata==4.8.3
mypy-extensions==1.0.0
mysqlclient==2.1.1
pathspec==0.9.0
platformdirs==2.4.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==1.2.3
typed-ast==1.5.4
typing_extensions==4.1.1
uWSGI==2.0.21
zipp==3.6.0
